common time mixin
